+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_models.py
@@ -7,7 +7,7 @@
 # pylint: disable=super-init-not-called, too-many-lines
 
 from azure.core.paging import PageIterator
-from ._parser import _parse_datetime_from_str, _datetime_to_str
+from ._parser import _parse_datetime_from_str
 from ._shared.response_handlers import return_context_and_deserialized, process_storage_error
 from ._shared.models import DictMixin, get_enum_value
 from ._generated.models import StorageErrorException
@@ -15,7 +15,6 @@ from ._generated.models import Metrics as GeneratedMetrics
 from ._generated.models import RetentionPolicy as GeneratedRetentionPolicy
 from ._generated.models import CorsRule as GeneratedCorsRule
 from ._generated.models import AccessPolicy as GenAccessPolicy
-from ._generated.models import CopyFileSmbInfo as GenCopyFileSmbInfo
 from ._generated.models import DirectoryItem
 
 
@@ -183,44 +182,6 @@ class AccessPolicy(GenAccessPolicy):
         self.start = start
         self.expiry = expiry
         self.permission = permission
-
-
-class CopyFileSmbInfo(GenCopyFileSmbInfo):
-    """Additional parameters for start_copy operation.
-
-    :param ignore_read_only: Specifies the option to overwrite the target file
-        if it already exists and has read-only attribute set.
-    :type ignore_read_only: bool
-    :param file_attributes: Specifies either the option to copy file
-        attributes from a source file(source) to a target file or a list of
-        attributes to set on a target file.
-        eg."source": copy file attributes from a source file(source),
-        "None": All attribute values will be cleared,
-        If this is not set, the default value is "Archive".
-    :paramtype file_attributes: str or :class:`~azure.storage.fileshare.NTFSAttributes`
-    :param file_creation_time: Specifies either the option to copy file
-        creation time from a source file(source) to a target file or a time value
-        in ISO 8601 format to set as creation time on a target file.
-        If this is not set, creation time will be set to the date time value of the creation(or when it was overwritten)
-        of the target file by copy engine.
-    :paramtype file_creation_time: str or ~datetime.datetime
-    :param file_last_write_time: Specifies either the option to copy file last
-        write time from a source file(source) to a target file or a time value in
-        ISO 8601 format to set as last write time on a target file.
-        If this is not set, value will be the last write time to the file by copy engine.
-    :paramtype file_last_write_time: str or ~datetime.datetime
-    :param set_archive_attribute: Specifies the option to set archive
-        attribute on a target file. True means archive attribute will be set on a
-        target file despite attribute overrides or a source file state.
-    :type set_archive_attribute: bool
-    """
-
-    def __init__(self, **kwargs):
-        self.ignore_read_only = kwargs.get('ignore_read_only', None)
-        self.file_attributes = kwargs.get('file_attributes', None)
-        self.file_creation_time = _datetime_to_str(kwargs.get('file_creation_time', ""))
-        self.file_last_write_time = _datetime_to_str(kwargs.get('file_last_write_time', ""))
-        self.set_archive_attribute = kwargs.get('set_archive_attribute', None)
 
 
 class LeaseProperties(DictMixin):